+++ mathesar_ui/src/pages/import/preview/ErrorInfo.svelte
