+++ mathesar_ui/src/pages/import/upload/DataFileInput.svelte
