+++ mathesar_ui/src/pages/import/upload/uploadUtils.ts
