+++ mathesar_ui/src/routes/ImportRoute.svelte
@@ -4,8 +4,8 @@
   import type { Database, SchemaEntry } from '@mathesar/AppTypes';
   import AppendBreadcrumb from '@mathesar/components/breadcrumb/AppendBreadcrumb.svelte';
   import { iconImportData } from '@mathesar/icons';
-  import ImportPreviewPage from '@mathesar/pages/import-preview/ImportPreviewPage.svelte';
-  import ImportUploadPage from '@mathesar/pages/import-upload/ImportUploadPage.svelte';
+  import ImportPreviewPage from '@mathesar/pages/import/preview/ImportPreviewPage.svelte';
+  import ImportUploadPage from '@mathesar/pages/import/upload/ImportUploadPage.svelte';
   import { getImportPageUrl, getImportPreviewPageQueryParams } from './urls';
 
   export let database: Database;
